Support passing environment variables to dependencies
Environment variables can be passed to the preBuild phase of
dependencies with the `-e` or `--env` cli argument.

Declare the argument multiple times to add multiple envrionment
variables.

This will add the environment variable to all generated packages. If
you only want to assign a variable to one package, use an overlay and
add a preBuild phase to that one package.